requirements.txt
object_detection
pandas
Pillow
tensorflow
